feat(utils): Создает отдельный файл для служебных функций. Переносит туда все лишнее из comments-model
